Omdøbt labels i Double80 assembler-kode med navn end til Exit (std-namespace er nu includeret og end giver derfor navnesammenfald) Bruger std::streamsize i operator<< for Rational og Double80

+++ Lib/Src/Math/Expression/AddentArray.cpp
@@ -48,7 +48,7 @@ void AddentArray::initDebugString() {
     m_debugString = _T("0");
   } else {
     m_debugString = (*this)[0]->getDebugString();
-    for(int i = 1; i < size(); i++) {
+    for(size_t i = 1; i < size(); i++) {
       const SumElement *e = (*this)[i];
       m_debugString += e->getDebugString();
     }
