+++ Lib/Src/Math/Expression/ExpressionNode.cpp
@@ -400,7 +400,7 @@ void ExpressionNodeArray::initDebugString() {
     m_debugString = _T("");
   } else {
     m_debugString = format(_T("(%s)"), (*this)[0]->getDebugString().cstr());
-    for(int i = 1; i < size(); i++) {
+    for(size_t i = 1; i < size(); i++) {
       m_debugString += format(_T(",(%s)"), (*this)[i]->getDebugString().cstr());
     }
   }
